+++ stories/Layout.stories.ts
@@ -1,0 +1,11 @@
+import Layout from './Layout.svelte';
+
+export default {
+  title: 'Design System/Components/Layout',
+  component: Layout,
+  parameters: {
+    layout: 'fullscreen'
+  }
+};
+
+export const PageStructure = {};